requirements.txt
Flask
# pyttsx3
wolframalpha
requests
wikipedia
gunicorn
python-dotenv
